Add chapter leader discovery module
